+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - lang: Remove the `interface-instructions` feature and the `#[interface]` attribute ([#4156](https://github.com/solana-foundation/anchor/pull/4156)).
 - cli: Remove the `login` command ([#4182](https://github.com/solana-foundation/anchor/pull/4182)).
 - idl: Exclude external accounts ([#4197](https://github.com/solana-foundation/anchor/pull/4197)).
+- idl: Remove the conflicting account names check ([#4294](https://github.com/solana-foundation/anchor/pull/4294)).
 
 ## [0.32.1] - 2025-10-09